testing

+++ src/redux/coupons/couponApis.js
@@ -8,7 +8,6 @@ export const getcoupons = async ({
   try {
     const { data } = await DataService.get(`/api/v1/coupons?page=${currentPage}&limit=${pageSize}`);
     setTotalPages(data.totalPages)
-    console.log(data);
     return data;
   } catch (error) {
     return error;
@@ -41,7 +40,7 @@ export const updatecoupon = async (data) => {
 };
 export const deletecoupon = async (body) => {
   try {
-    const { data } = await DataService.delete(`/api/v1/coupons/${body.id}`);
+    const { data } = await DataService.delete(`/api/v1/coupons/admin/${body.id}`);
     body.getData();
     return data;
   } catch (error) {
